add-implement hook index
