+++ src/components/auth/register/register-form.tsx
@@ -1,14 +1,11 @@
 'use client';
 import Spinner from '../../ui/spinner';
-import useGetGenders from '@/hooks/auth/register/useGetGenders';
-import useGetMaritalStatus from '@/hooks/auth/register/useGetMaritalStatus';
-import usePostEmployee from '@/hooks/auth/register/usePostEmployee';
-import useGetJobsPositions from '@/hooks/auth/register/useGetJobPositions';
-import useGetEmbargoes from '@/hooks/auth/register/useGetEmbargoes';
 import SelectField from '../../ui/select-fields';
 import InputField from '../../ui/input-field';
 import Link from 'next/link';
 import { Button } from '@/components/ui/button';
+import { useGetEmbargoes, useGetGenders, useGetMaritalStatus, usePostEmployee } from '@/hooks';
+import useGetJobsPositions from '@/hooks/auth/register/useGetJobPositions';
 
 const RegisterForm = () => {
   const { genders } = useGetGenders();
