+++ src/components/request-management/pay-slip/pay-slip-form.tsx
@@ -3,7 +3,7 @@ import { Button } from '@/components/ui/button';
 import InputField from '@/components/ui/input-field';
 import Spinner from '@/components/ui/spinner';
 import { titleFont } from '@/config/fonts';
-import usePostPaySlip from '@/hooks/request-management/pay-slip/usePostPaySlip';
+import { usePostPaySlip } from '@/hooks';
 
 const PaySlipForm = () => {
   const { onSubmit, register, mutation } = usePostPaySlip();
