+++ src/hooks/auth/change-password/usePostChangePassword.ts
@@ -3,11 +3,12 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation } from '@tanstack/react-query';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import toast from 'react-hot-toast';
-import { postChangePassword } from '@/server/auth/change-password/actions';
 import useGetToken from '@/hooks/common/useGetToken';
 import { z } from 'zod';
+
 import { ChangePassword } from '@/types';
 import { ChangePasswordSchema } from '@/schemas';
+import { postChangePassword } from '@/server';
 
 type FormsFields = z.infer<typeof ChangePasswordSchema>;
