+++ src/hooks/auth/register/useGetJobPositions.ts
@@ -1,5 +1,6 @@
-import { getJobsPositions } from '@/server/auth/register/actions';
 import { useQuery } from '@tanstack/react-query';
+
+import { getJobsPositions } from '@/server';
 
 const useGetJobsPositions = () => {
   const { data: JobsPositions } = useQuery({
