+++ src/hooks/auth/reset-password/usePostResetPasssword.ts
@@ -1,6 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
 import { SubmitHandler, useForm } from 'react-hook-form';
-import { postResetPassword } from '../../../server/auth/reset-password/actions';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useRouter } from 'next/navigation';
 import { signIn } from 'next-auth/react';
@@ -8,6 +7,7 @@ import toast from 'react-hot-toast';
 import { z } from 'zod';
 import { ResetPassword } from '@/types';
 import { ResetPasswordSchema } from '@/schemas';
+import { postResetPassword } from '@/server';
 
 interface Props {
   token: string;
