+++ src/hooks/index.ts
@@ -1,0 +1,24 @@
+export { default as useChangePassword } from './auth/change-password/usePostChangePassword';
+export { default as useForgotPassword } from './auth/forgot-password/usePostSendEmail';
+export { default as useLogin } from './auth/login/usePostLogin';
+export { default as useRegister } from './auth/register/usePostEmployee';
+export { default as usePostResetPassword } from './auth/reset-password/usePostResetPasssword';
+export { default as useGetEmbargoes } from './auth/register/useGetEmbargoes';
+export { default as useGetGenders } from './auth/register/useGetGenders';
+export { default as useGetJobPositions } from './auth/register/useGetJobPositions';
+export { default as useGetMaritalStatus } from './auth/register/useGetMaritalStatus';
+export { default as usePostEmployee } from './auth/register/usePostEmployee';
+
+
+export { default as useGetEmployeeId } from './common/useGetEmployeeId';
+export { default as useGetRoles } from './common/useGetRoles';
+export { default as useGetToken } from './common/useGetToken';
+
+export { default as useGetByIdEmployee } from './profile/useGetByIdEmployee';
+export { default as useUpdateEmployee } from './profile/useUpdateEmployee';
+
+export { default as usePostPaySlip } from './request-management/pay-slip/usePostPaySlip';
+export { default as usePostVacation } from './request-management/request-vacation/usePostVacation';
+export { default as usePostSalaryCetificates } from './request-management/salary-certificates/usePostSalaryCetificates';
+
+
