+++ src/hooks/profile/useGetByIdEmployee.ts
@@ -1,4 +1,4 @@
-import { getByIdEmployee } from '@/server/profile/getEmployee/actions';
+import { getByIdEmployee } from '@/server';
 import { useQuery } from '@tanstack/react-query';
 
 interface Props {
