+++ src/hooks/request-management/request-vacation/usePostVacation.ts
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import { useForm } from 'react-hook-form';
 import toast from 'react-hot-toast';
-import { postVacation } from '@/server/request-management/request-vacation/actions';
 import { RequestVacation } from '@/types';
+import { postVacation } from '@/server';
 
 const usePostVacation = () => {
   const { register, handleSubmit } = useForm<RequestVacation>();
